test: divided main and functions.cpp
